+++ src/StateR/Old/AsyncReducers/OperationStateUpdated.cs
@@ -1,4 +1,4 @@
-﻿namespace StateR
+﻿namespace StateR.Old
 {
     public record OperationStateUpdated (AsyncOperationState NewRecordState) : IAction;
 }
