+++ src/StateR/Old/BrowsableStates/BrowsableStateOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace StateR
+﻿namespace StateR.Old
 {
     public class BrowsableStateOptions
     {
